+++ api/src/main/java/marquez/logging/LoggingMdcFilter.java
@@ -1,0 +1,34 @@
+package marquez.logging;
+
+import java.io.IOException;
+import java.util.UUID;
+import javax.ws.rs.container.CompletionCallback;
+import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.container.ContainerRequestFilter;
+import org.slf4j.MDC;
+
+/**
+ * Puts request ID, method, and path to the MDC context so that this information is available from
+ * logs throughout the request context. Request ID is randomly generated UUID, which can be used to
+ * group by to see the logs for a particular request.
+ */
+public class LoggingMdcFilter implements ContainerRequestFilter, CompletionCallback {
+
+  private static final String REQUEST_ID = "requestID";
+  private static final String METHOD = "method";
+  private static final String PATH = "path";
+
+  @Override
+  public void onComplete(Throwable throwable) {
+    MDC.remove(REQUEST_ID);
+    MDC.remove(METHOD);
+    MDC.remove(PATH);
+  }
+
+  @Override
+  public void filter(ContainerRequestContext requestContext) throws IOException {
+    MDC.put(REQUEST_ID, UUID.randomUUID().toString());
+    MDC.put(METHOD, requestContext.getMethod());
+    MDC.put(PATH, requestContext.getUriInfo().getPath());
+  }
+}