+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/MarquezProject/marquez/compare/0.20.0...HEAD)
 
+### Added
+
+* Add MDC to the `LoggingMdcFilter` to include API method, path, and request ID.
+
 ## [0.20.0](https://github.com/MarquezProject/marquez/compare/0.19.1...0.20.0) - 2021-12-13
 
 ### Added